[FE-29] refactor : EnterPage 컴포넌트 분리
